+++ src/developer_options.py
@@ -16,7 +16,7 @@ class DeveloperOptions():
     """ Major options """
 
     # Version is represented as a string so we can graduate to x.y.z versioning in the near-future
-    version: str = "1.0.1"
+    version: str = "1.1.0"
 
     class ExecutionMode(Enum):
         # Standard internal release. Should override all debugging related options.
@@ -50,12 +50,21 @@ class DeveloperOptions():
     # The Qt Gui uses multi-threading to allow the interface to remain responsive while performing computationally
     # intensive work. Settings this flag to false forces LyricManager to be single-threaded, significantly easing debugging
     # and developing new code.
-    gui_multithreading_enabled      = True
-    #gui_multithreading_enabled      = False
+    #gui_multithreading_enabled      = True
+    gui_multithreading_enabled      = False
 
     @classmethod
     def is_multithreading_enabled(cls):
         return cls.is_release() or cls.gui_multithreading_enabled
+    
+
+    # With our limited resources, it's not feasible to maintain multiple schema versions. Thus, it's globally defined
+    # for the application, and when it is updated so must any software that relies upon it for now. :(
+    # MAJOR version when you make incompatible API changes,
+    # MINOR version when you add functionality in a backwards compatible manner, and
+    # PATCH version when you make backwards compatible bug fixes.
+    json_schema_version = "2.1.0"
+
     
     logging_level = logging.INFO
     #logging_level = logging.DEBUG
@@ -68,7 +77,7 @@ class DeveloperOptions():
             return cls.logging_level
 
 
-    model_execution_application =ModelExecutionApplication.Apptainer
+    model_execution_application = ModelExecutionApplication.Apptainer
 
     """ Minor options """